Fix nested labels.
Now when we enter a construct that
requires labels all labels for that
construct are allocated. These labels
are stored in a labels class of the
correct type. The stack is now composed
of these label classes (essentially a
container for all the labels for a given
construct). A variable is also kept
which is simply the last label of
emitted IR. This provides a reference
point for emitting future IR. Constructs
should only use this reference and never
manipulate the a labels class on the
stack except for the one they put on the stack.